+++ ElvUI_LuckyoneUI/Init.lua
@@ -46,12 +46,6 @@ local toons = {
 	["Luckydk-Everlook"]			= true, -- [Alliance] DK
 	["Luckykek-Everlook"]			= true, -- [Alliance] Shaman
 	["Luckyone-Giantstalker"]		= true, -- [Alliance] Paladin
-	-- Testing
-	["Luckyone-Koranos"]			= true, -- [Beta] Druid
-	["Kekleox-Tyrhold"]				= true, -- [Beta] Evoker
-	["Luckyevoker-Valdrakken"]		= true, -- [Beta] Evoker
-	["Luckyone-Valdrakken"]			= true, -- [Beta] Druid
-	["Luckyone-Nobundo"]			= true, -- [PTR] Druid
 }
 
 -- Constants